% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ryotaro.R
\name{theme_ryotaro}
\alias{theme_ryotaro}
\title{Ryotaro's ggplot2 theme}
\usage{
theme_ryotaro(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme.
}
\description{
Ryotaro's ggplot2 theme
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_ryotaro()
}
