% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_latex.R
\name{desc_latex}
\alias{desc_latex}
\title{Create a LaTeX descriptive statistics table}
\usage{
desc_latex(data, vars = NULL, digits = 2)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Character vector of variable names. If NULL, all numeric variables are used.}

\item{digits}{Number of digits to round numeric statistics.}
}
\value{
A LaTeX table produced by knitr::kable().
}
\description{
Create a LaTeX descriptive statistics table
}
\examples{
desc_latex(mtcars)
desc_latex(mtcars, vars = c("mpg", "hp", "wt"))
}
